Low: libcrmcommon: lower maximum IPC flush delay to 1.5 seconds
It appears that socket buffering behaves differently than the current pacemaker
IPC code expects. For example, a socket may return -EAGAIN for sending and
immediately be ready for another send. While a better approach will be
required, a simple workaround for now is to lower the flush delay. This will
help avoid some timeouts when sending large amounts of data.