Fix: scheduler: ensure orphaned recurring monitors have interval set
Normally, unpack_operation() sets a recurring monitor's interval in the action
meta-data based on the action XML. However, orphaned recurring monitors will
not have any XML. In that case, set the interval based on the operation key.

This fixes an issue where orphaned recurring monitors could not be cancelled
if the resource was unmanaged, because the action would not be detected as a
recurring monitor.