Refactor: fencing: use enum when creating device registration op
This hopefully makes the intent a bit clearer.

The previous cib_device_update() would attempt to pass the "provider" from the
fence device resource XML to create_device_registration_xml(). However, the
schema does not allow "provider" with "stonith"-class resources, so it would
always be NULL. Now it doesn't look for provider.