Fix: executor: cancel recurring monitors if fence device registration is lost
If the fence daemon crashes and respawns, any recurring monitors for fence
devices fail with "no device". Previously, this would be recorded as a monitor
failure, but the monitor would still be rescheduled indefinitely. Now, the
monitor will be cancelled in such a situation.
cancellation would fail because the fence device would no longer be registered.

Fixes CLBZ#5352