XML: resources-3.1.rng: relaxed schema after tight upgrade enforcement
As mentioned previously[*], the boundary, "transitional" version of
the schema, pacemaker-3.0, already serves as a barrier keenly enforcing
the deep upgrade process with the lower-versioned input, so now with
finalized upgrade process, it's about the time to relax some
restrictions again so that brand new configurations based on this
new upcoming minor version of the schema from the scratch are not
constrained merely because of otherwise artificial upgrade enforcers.

In particular, previous limitations on keys with instance_attributes,
fully application/agent specific data by definition, get now lifted
as those namespaces are no longer polluted with pacemaker's intimate
data (see also https://bugzilla.redhat.com/show_bug.cgi?id=1470223).
Similar limitations with meta_attributes are left in place, simply
because these are, in contrast, pacemaker-specific, and allowing
something that pacemaker won't be affected with (while it was in
the past) is rather counterproductive.

Also refresh expected "validity" subtest of cts-cli suite, as usual.

[*] see commits 14adc58d4 and 9ac70aa6b