Build: update shared object versions for 2.0.0 final release
This includes a one-time +20 bump of all public libraries' "current interface"
number. This will allow the 1.1 series to backport API changes from 2.0 and
use interface numbers in the intermediate range.

This is not the ideal way of handling the code split, but it should work fine
to keep applications from using incompatible shared objects. The 1.1 series
is not in active maintenance and will only get backports of 2.0 features, so
20 increments should be far more than enough. Since there are few users of the
pacemaker API, going to a lot of trouble for library versioning is not necessary.