Test: CTS: remove vestiges of former valgrind options
CTS supports the --valgrind-tests option to enable valgrind-specific tests,
and --valgrind-procs to specify which pacemaker daemons are running under
valgrind (to avoid killing them in a ComponentFail test, which wouldn't work
with the current design).

However the code also had further handling that no longer works. It checked
the value of the valgrind-prefix option when choosing valgrind log names, yet
didn't parse that option from the command line. Similarly, it used the value of
valgrind-opts when defining VALGRIND_OPTS, yet didn't parse that either. It set
a variety of environment variables intending to enable valgrind in pacemaker
when launching pacemaker, but since the start command is now "service start
corosync && service start pacemaker", the environment variables would only
apply to the first service execution (i.e. of corosync) and not to pacemaker,
and so were ineffective. As a further result, the supplied cts.supp valgrind
suppressions file, which has not been kept up to date and is superseded by
valgrind-pcmk.suppressions anyway, was not used. Also, since "service" is a
wrapper for systemctl in many distributions, any environment variables passed
to it would not be passed to pacemaker anyway.

Now, all that code is gone, and cts.supp is removed.

In the future, it may be worthwhile to bring back --valgrind-opts and
--valgrind-prefix, and enable valgrind via modifying /etc/sysconfig/pacemaker
on the cluster nodes.