Test: CTS: eliminate StandbyLoopTest
The StandbyLoopTest attempted to test for memory leaks by putting a node in and
out of standby mode for an hour. I am unsure why that was considered useful, but
the main test of memory leaks is running all the other tests while the daemons
are run by valgrind, so that as many code paths as possible are executed.

For now, there are no more valgrind-specific tests, however --valgrind-tests is
still useful to skip ComponentFail for daemons being run by valgrind.

This also disables find_leaks() in the StandbyLoopTest (which may still be run
explicitly), which should have been done with 0b221106, which removed logPat
thus causing find_leaks() to throw an exception.