Test: cts-scheduler: correct "origin" test expected output
The origin test is defined as:

    do_test origin "Timing of recurring operations" -t "2014-05-07 00:28:00"

Previously, the scheduler regression test built its crm_simulate command line
using $*, which means that the above was converted to a crm_simulate command
with:

    ... -S -t 2014-05-07 00:28:00

Since the -t argument wasn't quoted, only the first word (2014-05-07) was
actually used, and the 00:28:00 was treated as a regular positional argument
and ignored. The test command output thus was based on midnight 2014-05-07.

Now that the test correctly quotes its arguments, the originally intended time
is correctly used, thus changing the test output.