Low: scheduler: clear globally-unique correctly when forcing anonymous
Previously, when a globally-unique clone was forced to be anonymous, the
scheduler would recursively clear the parent's pe_rsc_unique flag. That had two
issues: the child being unpacked wasn't added to the parent's children yet, so
it wasn't cleared; and the globally-unique meta-attribute was not cleared.
Those issues are now fixed.