Test: scheduler: make inc6 regression test input more sane
The crm_simulate clone numbering change made the inc6 summary change.
However, inc6 has an impossible status section; one of the weird things was
that there were two separate lrm_resource history entries for child_rsc6:1 on
node2, one with a successful start, and one with a probe not-running result.

Since the live cluster would not handle this well either and should never have
to, remove the probe result, which makes the test output the same again.