Doc: tables: fix malformed column specifiers so as to please Asciidoctor
Documentations is clear about the order of the fields there:
- http://asciidoc.org/userguide.html#X70
- https://asciidoctor.org/docs/user-manual/#cols-format

While AsciiDoc was relaxed about the exact order, Asciidoctor is
apparently not and was turning affected columns as if the "literal"
style applied on them, which in turn raised validation errors, since
neither "emphasis" nor "xref" can be used within (with DocBook 4.5;
see the previous commit).

Also refrain from styling some, non-first table columns as "asciidoc"
(i.e., let them end up with the implicit "default" style), since it
precluded marking anchors for cross references, for instance.
On the other hand, "asciidoc" styling is a precondition for block
elements and source-only comments, so the affected cells needs
to be marked so specifically now.

Mention these new conventions in asciidoc.reference file for a clear
guidance on this slightly tricky topic.