Fix: pacemaker-based: inform originator of CIB upgrade failure
"cibadmin --upgrade" sends an upgrade request to its local CIB manager,
which forwards the request to the DC. If the DC successfully upgrades the CIB,
it broadcasts a message causing all peers to perform the upgrade, which also
results in local clients such as cibadmin being notified.

However, if the upgrade is either not necessary or fails, the DC would simply
ignore the request, and local clients would time out waiting for a reply that
would never come. Now, the DC will send the error result to the originating
peer, which will notify its clients.