Fix: scheduler: Act on clone instances in numerical order
Previously, ordered clone instances were ordered alphabetically, which meant
that instance :10 came between :1 and :2.

This does not cause any serious problems, but a user who configures a clone to
be both unique and ordered may reasonably expect that instances will be started
in numerical order.

A separate question (not addressed here) is whether it would be reasonable to
expect ordered instances to be stopped in reverse numerical order.