Build: Makefile.common: avoid using verbatim control characters
This is not a good practise, since it poses an imminent risk of
"deceptive display (rendering)" at various places, especially
tools that may be used for change reviews, incl. GitHub[1].

When at it, arrange for periodical CI to capture undesired
occurrences of such characters in the code base, and related
to that, obtain pristine LGPLv2.1 license text anew using

  curl  https://www.gnu.org/licenses/lgpl-2.1.txt \
    | tr -d '\f' > licenses/LGPLv2.1

so that embedded "form feed" characters are got rid of (beside
that, also tabulators are changed into spaces, which is how this
authoritative server currently carries it).

[1] https://github.com/isaacs/github/issues/1329