Build: autogen.sh: get back to bare-bones, do not autotrigger configure
It was needlessly bloated with an exteremly limited usefulness.

Regarding the latter, it is not a customary practice (first
project out of numerous ones I witnessed, actually), depends on
an unpublished implementation detail of autoconf, and is also
counter-productive when one changes something in configure.ac and
only wants to test the configure options are available towards the
user afterwards.

Since a one-liner is not by any means anything of the originality,
also drop the original license notice, which would then simply
amount to another ballast, as the whole inventive content there
was to be covered exists no more.  Instead, use the convention to
associate file with the project and add an SPDX license tag
referencing the project's default license as of the moment.

Also regarding GREP_OPTIONS= workaround, note that it was in most
likeliness needed only for very ancient autoconf versions, which
got fixed there in 5f19cfb8 (meaning it's no longer needed for
v2.63b+, which is already mandated by configure.ac proper
[with AC_PREREQ(2.64)]).