Build: libcrmcommon: explicitly include stdint.h in util.h
Required as of bdf69722 which added a reference to uint32_t.

Until now, all use of uint32_t and similar has indirectly relied on getting
them from crm/common/logging.h, which includes qb/qblog.h, which includes
stdint.h. As everything includes util.h one way or another, this should be
sufficient to make it explicit within pacemaker.

This does not do any replacement of stdint.h for systems that do not have it,
since libqb didn't either, so we've already been relying on it indirectly.