High: stonith-ng's function cannot be blocked with CIB updates forever
In the high-load (or high-rate-config-change) scenarios,
pacemaker-fenced would be unable to provide service when basically DoS'd
with CIB update notifications.  Try to reconcile that with elevated
priority of the server's proper listening interface in the mainloop, at
worst, it will try to fence with slightly outdated config, but appears
to be less bad than not carrying the execution at all, for instance.
Other daemons might be considered as well.

Prerequisites:
- https://github.com/ClusterLabs/libqb/pull/352
  (libqb used to contain a bug due to which one particular step in the
  initial-client-connection-accepting-at-the-server procedure that would
  be carried out with hard-coded (and hence possibly lower than competing
  events') priority, which backfires exactly in this case (once the
  pacemaker part is fixed -- by the means of elevating priority for
  the API end-point of fenced so that it won't get consistently
  overridden with a non-socket-based event source/trigger)

How to verify:
- mocked/based -N (see commit adding that module to mocked based daemon)