Fix: libcrmcluster: reset "when lost" when a peer becomes a member
We track when a peer's CPG membership diverges from its cluster membership, and
if that lasts for more than a minute, we update its state to member.

Previously, we reset the tracking only if we updated the state to member at
that point. If the state was updated to member via other means, the tracking
timer would continue ticking, so if the peer left and rejoined the cluster
again, it would wrongly appear to have been divergent for more than a minute.

Now, the tracking is reset anytime a peer's state is updated to member.