Fix: controller: trigger join after both peer and join notifications
Previously, the controller's peer change callback would register an I_NODE_JOIN
only if the peer notification was received after the quorum notification.
However, this order is not guaranteed.

Now, the callback registers I_NODE_JOIN after both the quorum and peer join
notifications have been received for a node, regardless of the order in which
they occur.