Maint: gdbhelpers: start (finally) making our debugging work easier
Set forth on the systematic tracking of what are seemingly one-off
helpers for debugging purposes, since odds are it will be useful also
in the future.  Hopefully it will be kept in sync with any releavant
code-wise changes.

Once it will get richer on what it offers, it would be sensible to ship
that in the distribution (as in "make dist") and consequently in distro
specific (sub-)package, like already happens e.g. with librevenge-gdb in
Fedora.  For the time being, running gdb with

  -x <pcmk-checkout>/extra/gdb/gdbhelpers

must suffice.  Still can be a tangible effectivity booster in some
cases -- a single one now to be precise, but more to come as needed.