Low: libfencing: enumerate installed RHCS-style agents alpha-sorted
... and not in the reverse thereof -- reported by Tomáš Jelínek
(of Red Hat).

Also ensure that full capacity of FILENAME_MAX incl. the terminating
null character can be utilized (or whatever size the respective buffer
is defined with, no lessened by one for no reason); on that note, C99
doesn't seem to provide clear guidance whether the trailing null is
assumed and counted in, so current extra +1 offset is a conservative
measure when it's not, since we still want to pass null-terminated
string down the line while maintaining the full capacity compatibility
(that glibc manual warns that it may be arbitrarily big and hence
should not be used for defining static character arrays is a different
story, trivially solvable when POSIX rev. 2008+ is applicable, which
we apply here as well -- sadly, fstatat gnulib module is covered with
the terms of GPL rather than LGPL that's needed for compatibility).