Low: attrd: record tentative writer only if election not already lost
Previously, we could lose an election to another node and tentatively set it as
writer, then get a no-vote from another node and reset writer to it because we
had already lost the election. The proper winner would eventually be learned,
but it's better to avoid tentatively setting a node as writer that we know
won't win.