Low: attrd: connect to the CIB before connecting the cluster
This avoids a (highly unlikely) race: If the local node wins a writer election
and receives a peer update at start-up, between the cluster connection and the
CIB connection, we could delay the write due to "cib not connected". However,
we would not re-attempt this write once the connection was established, so the
value might never be written out (or delayed an unreasonably long time).

Rearranging the connections is a simple solution, and also allows us to assume
the CIB is connected everywhere else.