Low: controller: remove node's vote as soon as it is inactive
Previously, a lost peer's vote (if any) in the DC election would be removed
when reaping dead nodes as part of the corosync quorum callback. However,
election votes are counted against the number of active peers, so it is
important to remove a vote whenever a peer would evaluate as inactive.

Now, the vote is removed in the peer status callback if either CPG
membership or cluster membership is lost, either of which causes a node to
evaluate as inactive.