Low: scheduler: avoid container stop on wrong node
Previously, a failed resource with an on-fail of restart-container would lead
to a stop of the container on the failed resource's node. However, a bundle's
connection can run on a different node than its container, so that this could
be the wrong node. Now, remember that the container needs to be stopped, and
stop it once its location is known.