Low: libcrmcommon: allow for fail-safe user translation ACL arrangement
This is a follow-up for 186edf35a as subsequently discussed at [1]
and summarized in the in-code comment.  It's merely an opt-in
opportunity and should be harmless, because once some entity is in
charge of the user database, all bets are off, and ACLs are fully in
scope of the administrators (the same story, actually, once the less
privileged is allowed to mangle with in-CIB ACLs by the means of
pre-arranged ACLs, it's consequently an ineffective measure, so only
the self-inflicted escalations are the [self-understood, hopefully]
risk here).

[1] https://github.com/ClusterLabs/pacemaker/pull/1563/commits/186edf35a47372ae0e30117d9e18c312290a7381#diff-f0c579bcdba77d0ea4c6f0a89ee91b86R305