Fix: scheduler: order guest pseudo-fencing properly after clean-up
If the resource history of a guest node's container has been cleaned, we will
schedule a (pseudo-)fence of the guest node, and a stop of the guest node's
connection resource, but not a stop of the container (which appears already
stopped). In this case, order the pseudo-fence after the connection stop, so we
don't call remote_node_down() unless the container is really down (the
connection stop will be avoided if the container is really up).