Feature: libpe_status: add constructor/destructor for pe_working_set_t
The data type has been problematic up to this point, since callers typically
declare it statically. That means that adding new members to the end of the
structure breaks API backward compatibility. The API now mandates that this
constructor be used to create instances, to avoid this issue.

Since the API has not yet been documented, we can consider it not yet
officially released until this point, so there technically is no public API to
break with any structure changes. But in practice such changes would have the
potential to break any use of pe_working_set_t objects without these functions.