Log: attrd: clear lost updates
If an attribute's CIB update is determined to be lost, set the update ID to 0,
to avoid logging the same message repeatedly in the (unlikely) chance that the
a new write couldn't be scheduled (e.g. if the peer UUIDs are unknown for the
values that need to be written).