Refactor: libpe_status: get rid of pe_dataset global
There was never a guarantee or documented requirement that the calling
application would keep the working set instance in memory for the entire time
it would be needed via the global, though it would in practice.

As long as the new constructor/destructor functions are used for a single
active working set instance, it still shouldn't ever be a problem, but it seems
more user-friendly to allow the possibility of multiple instances.