Fix: fenced: adhere meta-attributes when triggering device-update
Revert previous fix that worked via some dirt-effect but triggered
much too often on status-updates. When pcs enables a fence-resource
it deletes the target-role that was mistakenly interpreted as
device deletion. Attempt to do low effort analysis of diffs is
always a tradeoff between effort and risk of missing certain
cases. (default-target-role as an example for something still
uncovered)