Fix: scheduler: order resource moves after connection starts
This addresses a regression in behavior since 1.1.18 (via 3a34fed). By allowing
stops to proceed before probes finished, that change allowed the stop of a
resource moving to a coming-up remote node to happen before the remote node
connection's start. If the remote connection start fails, the resource will
have to be started again where it was, leading to unnecessary downtime.

Now, order the resource's stop after the remote connection's start.

RHBZ#1648507