Low: scheduler: order start after particular stonith op
Previously, if a resource's status was unknown on a node about to be fenced,
we ordered the resource start after all_stopped. This approximated stonith_done
before that was available. However, it makes more sense to order the start
after the particular stonith op for the node in question.

This improves recovery when multiple nodes are being fenced: resources can now
be recovered from one node when it is successfully fenced, even if the fencing
of another node fails.