Feature: tools: Add crm_rule.
This is a new tool that can be used to investigate rules in the CIB.For
the moment, it only allows checking if a rule is expired or not, based
on its lifetime.  It's also a little more limited than that - due to
complexities in evaluating rules, it only handles checking rules that
have a single date_expression.

I expect more stuff to be added to this program in the future, which is
why I have over-written some of the command line handling code.