Feature: tools: crm_resource --clear should print out what it does
This is accomplished in the most obvious, and probably least clean way
possible.  A list of constraints is generated before clearing, another
list of constraints is generated after clearing, and the difference is
taken.  Anything in the difference was cleared.  Print all those out.

Also, update the CLI regression test case.

See rhbz#1631752